transcendance/requirements.txt
django
djangorestframework==3.12.4
dj-rest-auth